+++ test/MicroElements.Swashbuckle.FluentValidation.Tests/SchemaGenerationTests.cs
@@ -744,5 +744,6 @@ namespace MicroElements.Swashbuckle.FluentValidation.Tests
                 }
             }
         }
+
     }
 }